connect_mcp_server/pyproject.toml
[project]
name = "connect-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp-server-time>=0.6.2",
    "openai-agents>=0.0.15",
]
